Revert "Correct various types warnings"
This reverts commit d415dba9a6bcad6aa27dd791aa023b97cbd7c91b.

+++ include/frame.h
@@ -19,7 +19,7 @@
 struct vector_frame;
 
 struct frame {
-    size_t size;
+    unsigned short size;
     void *frame;
     void *raw_packet;
 };
